TextGridShiny {TextGridShiny}	R Documentation

Annotation Interface for Speech Corpus

Description

TextGridShiny allows to read TextGrid files and to display the annotated data. In addition, it is possible to interact with the created visual annotation in adding tiers and labels.

Usage

TextGridShiny(message, width = NULL, height = NULL, elementId = NULL)

Arguments

message
A textgrid file containing the labels.

width
Width of the resulting graph - defaults to fit container, probably leave this alone

height
Height of the resulting graph - defaults to fit container, probably leave this alone

elementId
Do not specify, this is used by the 'htmlwidgets' package

Value

A Annotation Interface which is an 'htmlwidgets' object

Examples

## Not run: 

library(TextGridShiny)

TextGridShiny("inst/extdata/oran.TextGrid")

## End(Not run)

[Package TextGridShiny version 1.0.0 Index]