TextGridShiny-shiny {TextGridShiny}	R Documentation

Shiny bindings for TextGridShiny

Description

Output and render functions for using TextGridShiny within Shiny applications and interactive Rmd documents.

Usage

TextGridShinyOutput(outputId, width = "100%", height = "100%")

renderTextGridShiny(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId
output variable to read from

width, height
Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

expr
An expression that generates a TextGridShiny

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

Examples

#----------------------- Most basic example -----------------

## Not run: 

library(shiny)

library(TextGridShiny)

app <- shinyApp(

  ui = fluidPage(
	TextGridShinyOutput("mytextgridannotation")
	),

 server = function(input, output) {
    output$mytextgridannotation <- renderTextGridShiny(
      TextGridShiny("inst/extdata/oran.TextGrid")
    )
  }
)

runApp(app)

## End(Not run)

[Package TextGridShiny version 1.0.0 Index]